add stapling-multi-file patch
This makes ssl_stapling_file an array of one file per certificate
for the ECDSA+RSA(+future) multi-certificate case.  In general, it
gives a fatal error if ssl_stapling_file is used without one such
file per certificate.

However, a backwards-compatible exception is made: if multiple
certificates are configured, but only a single ssl_stapling_file,
the emerge-level error is still logged, but configuration is
allowed to proceed and the filename is copied around to use for
all certificates, emulating the previous behavior.  Without this
it's hard to get through an upgrade as there would be no single
configuration file containing ssl_stapling_file that's compatible
across this commit.

Bug: T93927
Change-Id: Ie076b36732d3147262a5054bd8f2714179e34248